refactor: 일정 시간 후 보이지 않게 하는 기능 커스텀 훅으로 변경
