+++ src/components/Alert/index.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import useVisibleTemporarily from '@Hooks/useVisibleTemporarily';
 
 import * as S from './style';
 import * as T from './type';
@@ -9,15 +9,7 @@ function Alert({
   duration = 3,
   isReverse = false,
 }: T.Alert) {
-  const [isVisible, setIsVisible] = useState(true);
-
-  useEffect(() => {
-    const timeout = setTimeout(() => {
-      setIsVisible(false);
-    }, duration * 1000);
-
-    return () => clearTimeout(timeout);
-  }, []);
+  const isVisible = useVisibleTemporarily(duration);
 
   return isVisible ? (
     <S.Alert type={type} isReverse={isReverse}>
